chore(log,run): changed option to show full run details by default
Until projects have the ability to configure the display of run details,
we've made the decision to show full run details as the default
behavior.
This ensures comprehensive information is readily available to users.

This change is temporary and will be revisited once we implement
project-specific configuration options for run detail display.