fix(build): correct proxying to derive API path from URL_PREFIX
Previously, the BASE_URL environment variable was incorrectly used
to derive the API prefix path in development. The correct variable
to use is URL_PREFIX as defined in the environment example file.

This fix ensures we use URL_PREFIX and strip any `/v2` suffix
when needed, providing a more accurate development proxy setup.

Signed-off-by: Danil Kostromin <danil.kostromin@arknetworks.am>